refactor(imports): consolidate imports of ScreenplayUtils and get_default_llm_client
Changed multiple import statements to import ScreenplayUtils and get_default_llm_client directly from scriptrag.utils instead of scriptrag.utils.screenplay or separate imports. This simplifies and standardizes import paths across several modules and tests.
